+++ src/analyze-action.ts
@@ -106,6 +106,8 @@ async function run() {
     const memory = util.getMemoryFlag(
       actionsUtil.getOptionalInput("ram") || process.env["CODEQL_RAM"]
     );
+    const optResultLimit = actionsUtil.getOptionalInput("results-limit") || process.env["CODEQL_RESULTS_LIMIT"]
+    const resultsLimit = (optResultLimit) ? Number(optResultLimit) : Number.MAX_SAFE_INTEGER;
 
     const repositoryNwo = parseRepositoryNwo(
       util.getRequiredEnvParam("GITHUB_REPOSITORY")
@@ -185,6 +187,7 @@ async function run() {
         outputDir,
         config.gitHubVersion,
         apiDetails,
+        resultsLimit,
         logger
       );
       core.setOutput("sarif-id", uploadResult.sarifID);